+++ example/lib/pages/home_page.dart
@@ -24,6 +24,7 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     Get.put(PayPageController());
+
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.black,