+++ lib/motor_flutter.dart
@@ -1,3 +1,14 @@
+/// ## Motor Flutter
+///
+/// **Flutter bindings for the All-in-one Motor node by Sonr. Seamless integration with the Sonr P2P network and blockchain - all while providing a Wallet interface which doesnt require a Seed phrase.**
+/// - [Motor API Reference](https://motor.build/motor_flutter/MotorFlutter-class.html)
+/// - [Documentation](https://docs.sonr.io)
+/// - [Discord](https://sonr.io/discord)
+/// - [Website](https://sonr.io)
 library motor_flutter;
 
 export 'src/motor_flutter_base.dart';
+export 'src/types/types.dart';
+export 'src/utilities/logger.dart';
+export 'src/widgets/widgets.dart';
+export 'src/controllers/register_controller.dart';