+++ lib/src/widgets/register_button.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:motor_flutter/motor_flutter.dart';
 
 class SonrRegisterButton extends StatelessWidget {
@@ -12,20 +11,26 @@ class SonrRegisterButton extends StatelessWidget {
     return ButtonTheme(
       minWidth: 300.0,
       child: MaterialButton(
-        padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 24),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(16.0),
-        ),
-        color: Colors.black,
-        child: const Text(
-          'Register Account',
-          style: TextStyle(color: Colors.white, fontSize: 18),
-        ),
-        onPressed: () => Get.dialog(RegisterModalPage(
-          onCreateAccountResponse: onSuccess,
-          onError: onError,
-        )),
-      ),
+          padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 24),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(16.0),
+          ),
+          color: Colors.black,
+          child: const Text(
+            'Register Account',
+            style: TextStyle(color: Colors.white, fontSize: 18),
+          ),
+          onPressed: () async {
+            final res = await MotorFlutter.to.showRegisterModal(
+              onError: onError,
+            );
+            if (res == null) {
+              return;
+            }
+            if (onSuccess != null) {
+              onSuccess!(res);
+            }
+          }),
     );
   }
 }